Move button definition
