+++ ubitx_ui.cpp
@@ -1,4 +1,5 @@
 #include <Arduino.h>
+#include "button.h"
 #include "morse.h"
 #include "nano_gui.h"
 #include "settings.h"
@@ -77,14 +78,6 @@ enum btn_set_e {
   BUTTON_MNU,
   BUTTON_FRQ,
   BUTTON_TOTAL
-};
-
-struct Button {
-  int x, y, w, h;
-  unsigned int id;
-  char text[5];
-  char morse;
-  void (*morse_status)(int8_t* val_out);//-1 if a low tone should play, +1 if a high tone should play
 };
 
 void msVfoA(int8_t* val_out){
